move EPUB parser to epub package
